# docs/openapi/mediflow-v1.yaml
openapi: 3.1.0
info:
  title: MediFlow Local API v1
  version: 1.16.0
  summary: Stable patient contract plus executable home-base network slice with explicit pairing review, paired-client auth boundary, ambulatory scope options, scoped remote patient, diary, therapy, checkup, observation, and catalog read boundaries, replica planning state, explicit identity summary, and declarative AI runtime summary for the local `/api/v1` surface.
  description: |
    OpenAPI contract for the client-facing local API defined by ADR 0010.
    Current thin slice covers stable patient reads plus compare-on-write patient
    mutations on `PUT /api/v1/patients/{id}` and `DELETE /api/v1/patients/{id}`.
    It also introduces the first documented `/api/v1/network/*` contract slice
    for home-base discovery, session gating, capability discovery, explicit
    pairing bootstrap/review on trusted local networks, the first read-only
    remote patient data plane, the first paired patient profile/status write
    boundary, the first paired clinical diary read/create/update boundary, the
    first paired therapy read/create/update/soft-delete boundary, the first
    paired checkup read/create/update/soft-delete boundary, the first paired
    observation read/create/update/soft-delete boundary, and the first paired
    read-only catalog boundary for global drug, exemption, and terminology
    reference data behind dedicated paired-client auth capabilities.
...
servers:
  - url: https://127.0.0.1:3443
...
paths:
...
  /api/v1/network/terminology/search:
    get:
      tags:
        - Network
      operationId: searchNetworkTerminology
      summary: Search terminology catalog entries through the paired home-base data plane
      description: |
        Returns the same global terminology search shape exposed by the host
        `/api/v1/terminology/search` read route. This paired route requires
        the `network.catalogs.readonly` capability plus a valid paired client
        credential and operator session cookie. Terminology catalog data is
        global reference data and is not filtered by ambulatory scope.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      parameters:
        - name: system
          in: query
          required: true
          description: Terminology system to search.
          schema:
            $ref: '#/components/schemas/TerminologySystemCode'
        - name: q
          in: query
          required: false
          description: Optional search text.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Maximum number of rows to return.
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        '200':
          description: Matching terminology entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TerminologyItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/network/terminology/resolve:
    get:
      tags:
        - Network
      operationId: resolveNetworkTerminology
      summary: Resolve a terminology code through the paired home-base data plane
      description: |
        Returns the same global terminology resolve shape exposed by the host
        `/api/v1/terminology/resolve` read route. This paired route requires
        the `network.catalogs.readonly` capability plus a valid paired client
        credential and operator session cookie. Terminology catalog data is
        global reference data and is not filtered by ambulatory scope.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      parameters:
        - name: system
          in: query
          required: true
          description: Terminology system to resolve.
          schema:
            $ref: '#/components/schemas/TerminologySystemCode'
        - name: code
          in: query
          required: true
          description: Code to resolve within the requested terminology system.
          schema:
            type: string
      responses:
        '200':
          description: Resolved terminology entry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminologyItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/network/service-prescriptions:
    get:
      tags:
        - Network
      operationId: listNetworkServicePrescriptions
      summary: List service prescriptions through the paired home-base data plane
      description: |
        Lists service prescriptions for a patient in the paired client's
        effective ambulatory scope. This route requires
        `network.replica.readonly-service-prescriptions`.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      parameters:
        - name: patientId
          in: query
          required: true
          description: Patient identifier scoped to the paired client's ambulatory.
          schema:
            type: string
      responses:
        '200':
          description: Scoped service prescriptions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServicePrescriptionDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Network
      operationId: createNetworkServicePrescription
      summary: Create service prescription through the paired boundary
      description: |
        Creates a service prescription for a patient in the paired client's
        effective ambulatory scope. This route requires
        `network.replica.write-service-prescriptions`.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServicePrescriptionCreateRequest'
      responses:
        '201':
          description: Created service prescription identifier and version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedVersionedResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/network/service-prescriptions/{id}:
    put:
      tags:
        - Network
      operationId: updateNetworkServicePrescription
      summary: Update service prescription through the paired boundary
      description: |
        Applies a compare-on-write update to a service prescription in scope.
        `version` is required and stale versions return `VERSION_CONFLICT`.
        This route requires `network.replica.write-service-prescriptions`.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Service prescription identifier.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServicePrescriptionMutationRequest'
      responses:
        '200':
          description: Service prescription updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/PrescriptionVersionConflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/network/service-prescription-items/{id}:
    put:
      tags:
        - Network
      operationId: updateNetworkServicePrescriptionItem
      summary: Update service prescription item through the paired boundary
      description: |
        Applies a compare-on-write update to a service prescription item in
        scope. `version` is required and stale versions return
        `VERSION_CONFLICT`. This route requires
        `network.replica.write-service-prescriptions`.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Service prescription item identifier.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServicePrescriptionItemMutationRequest'
      responses:
        '200':
          description: Service prescription item updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/PrescriptionVersionConflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/network/prosthetic-prescriptions:
    get:
      tags:
        - Network
      operationId: listNetworkProstheticPrescriptions
      summary: List prosthetic prescriptions through the paired home-base data plane
      description: |
        Lists prosthetic prescriptions for a patient in the paired client's
        effective ambulatory scope. This route requires
        `network.replica.readonly-prosthetic-prescriptions`.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      parameters:
        - name: patientId
          in: query
          required: true
          description: Patient identifier scoped to the paired client's ambulatory.
          schema:
            type: string
      responses:
        '200':
          description: Scoped prosthetic prescriptions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProstheticPrescriptionDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Network
      operationId: createNetworkProstheticPrescription
      summary: Create prosthetic prescription through the paired boundary
      description: |
        Creates a prosthetic prescription for a patient in the paired client's
        effective ambulatory scope. This route requires
        `network.replica.write-prosthetic-prescriptions`.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProstheticPrescriptionCreateRequest'
      responses:
        '201':
          description: Created prosthetic prescription identifier and version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedVersionedResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
      description: "Local API token expected in `Authorization: Bearer <token>`."
    pairedClientId:
      type: apiKey
      in: header
      name: x-mediflow-paired-client-id
      description: Explicit paired-device identifier minted during local pairing confirmation.
    pairedClientToken:
      type: apiKey
      in: header
      name: x-mediflow-paired-client-token
      description: One-time paired-device token minted during local pairing confirmation.
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: mediflow_session
      description: Operator session cookie required for the remote network data plane.
  responses:
    BadRequest:
      description: Request payload is invalid or missing required mutation metadata.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingVersion:
              value:
                error: Version is required
            noFields:
              value:
                error: No valid fields to update
...
    NetworkDataPlaneUnauthorized:
      description: Missing or invalid paired client credential, or missing operator session cookie.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                error: Unauthorized
    Forbidden:
      description: Request is authenticated but blocked by missing scope or missing granted capability.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                error: Forbidden
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                error: Not found
...
    PrescriptionVersionConflict:
      description: |
        Prescription compare-on-write conflict. The payload exposes only
        PHI-safe metadata so clients can refresh and decide whether to retry.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PrescriptionVersionConflictResponse'
          examples:
            outdatedVersion:
              value:
                error: Conflict
                code: VERSION_CONFLICT
                entity: service_prescription
                recordId: prescription_123
                expectedVersion: 2
                currentVersion: 3
                currentUpdatedAt: '2026-05-02T10:15:00.000Z'
                currentState: present
                currentSnapshot:
                  id: prescription_123
                  patientId: patient_123
                  version: 3
                  updatedAt: '2026-05-02T10:15:00.000Z'
            missingPrescription:
              value:
                error: Conflict
                code: VERSION_CONFLICT
                entity: service_prescription
                recordId: prescription_123
                expectedVersion: 2
                currentVersion: null
                currentUpdatedAt: null
                currentState: missing
                currentSnapshot: null
    InternalServerError:
      description: Unexpected server-side failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            patientListFailure:
              value:
                error: Failed to fetch patients
            patientDetailFailure:
              value:
                error: Failed to fetch patient
  schemas:
...
    TerminologySystemCode:
      type: string
      enum:
        - AIC
        - ATC
        - ICD-11
        - LOINC
        - UCUM
        - SNOMED-CT
        - CND
    TerminologyItem:
      type: object
      additionalProperties: false
      required:
        - system
        - code
        - display
        - source
      properties:
        system:
          $ref: '#/components/schemas/TerminologySystemCode'
        code:
          type: string
        display:
          type: string
        displayIt:
          type: string
        defaultUnit:
          type: string
        version:
          type:
            - string
            - 'null'
        source:
          type: string
...
    CreatedVersionedResource:
      type: object
      additionalProperties: false
      required:
        - id
        - version
      properties:
        id:
          type: string
        version:
          type: integer
          minimum: 1
    MutationSuccessResponse:
      type: object
      additionalProperties: false
      required:
        - success
      properties:
        success:
          type: boolean
          const: true
    ServicePrescriptionDetail:
      type: object
      additionalProperties: false
      required:
        - id
        - patientId
        - prescribedAt
        - status
        - category
        - serviceName
        - version
      properties:
        id:
          type: string
        patientId:
          type: string
        prescribedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [prescribed, booked, performed, report_received, cancelled]
        category:
          type: string
          enum: [lab, imaging, visit, rehab, screening, procedure, other]
        priority:
          type:
            - string
            - 'null'
          enum: [U, B, D, P, routine, unknown, null]
        codeSystem:
          type:
            - string
            - 'null'
        serviceCode:
          type:
            - string
            - 'null'
        serviceName:
          type: string
        clinicalQuestion:
          type:
            - string
            - 'null'
        provider:
          type:
            - string
            - 'null'
        scheduledAt:
          type:
            - string
            - 'null'
          format: date-time
        performedAt:
          type:
            - string
            - 'null'
          format: date-time
        reportReceivedAt:
          type:
            - string
            - 'null'
          format: date-time
        outcomeNote:
          type:
            - string
            - 'null'
        requestReference:
          type:
            - string
            - 'null'
        source:
          type: string
          enum: [manual, document_review, legacy_therapy_cleanup]
        documentRefs:
          type:
            - string
            - 'null'
        notes:
          type:
            - string
            - 'null'
        version:
          type: integer
          minimum: 1
        createdAt:
          type:
            - string
            - 'null'
          format: date-time
        updatedAt:
          type:
            - string
            - 'null'
          format: date-time
    ServicePrescriptionCreateRequest:
      type: object
      additionalProperties: false
      required:
        - patientId
        - prescribedAt
        - serviceName
      properties:
        id:
          type: string
        patientId:
          type: string
        prescribedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [prescribed, booked, performed, report_received, cancelled]
        category:
          type: string
          enum: [lab, imaging, visit, rehab, screening, procedure, other]
        priority:
          type:
            - string
            - 'null'
          enum: [U, B, D, P, routine, unknown, null]
        codeSystem:
          type:
            - string
            - 'null'
        serviceCode:
          type:
            - string
            - 'null'
        serviceName:
          type: string
        clinicalQuestion:
          type:
            - string
            - 'null'
        provider:
          type:
            - string
            - 'null'
        scheduledAt:
          type:
            - string
            - 'null'
          format: date-time
        performedAt:
          type:
            - string
            - 'null'
          format: date-time
        reportReceivedAt:
          type:
            - string
            - 'null'
          format: date-time
        outcomeNote:
          type:
            - string
            - 'null'
        requestReference:
          type:
            - string
            - 'null'
        source:
          type: string
          enum: [manual, document_review, legacy_therapy_cleanup]
        documentRefs:
          type:
            - string
            - 'null'
        notes:
          type:
            - string
            - 'null'
    ServicePrescriptionMutationRequest:
      type: object
      additionalProperties: false
      required:
        - version
      properties:
        version:
          type: integer
          minimum: 1
        prescribedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [prescribed, booked, performed, report_received, cancelled]
        category:
          type: string
          enum: [lab, imaging, visit, rehab, screening, procedure, other]
        priority:
          type:
            - string
            - 'null'
          enum: [U, B, D, P, routine, unknown, null]
        codeSystem:
          type:
            - string
            - 'null'
        serviceCode:
          type:
            - string
            - 'null'
        serviceName:
          type: string
        clinicalQuestion:
          type:
            - string
            - 'null'
        provider:
          type:
            - string
            - 'null'
        scheduledAt:
          type:
            - string
            - 'null'
          format: date-time
        performedAt:
          type:
            - string
            - 'null'
          format: date-time
        reportReceivedAt:
          type:
            - string
            - 'null'
          format: date-time
        outcomeNote:
          type:
            - string
            - 'null'
        requestReference:
          type:
            - string
            - 'null'
        source:
          type: string
          enum: [manual, document_review, legacy_therapy_cleanup]
        documentRefs:
          type:
            - string
            - 'null'
        notes:
          type:
            - string
            - 'null'
...
    ServicePrescriptionItemMutationRequest:
      type: object
      additionalProperties: false
      required:
        - version
      properties:
        version:
          type: integer
          minimum: 1
        ordinal:
          type: integer
        status:
          type: string
          enum: [prescribed, booked, performed, report_received, cancelled]
        category:
          type:
            - string
            - 'null'
          enum: [lab, imaging, visit, rehab, screening, procedure, other, null]
        codeSystem:
          type:
            - string
            - 'null'
        serviceCode:
          type:
            - string
            - 'null'
        serviceName:
          type: string
        catalogEntryId:
          type:
            - string
            - 'null'
        catalogDisplayName:
          type:
            - string
            - 'null'
        matchStatus:
          type: string
          enum: [unmatched, candidate, matched, manual, not_found]
        confidence:
          type:
            - string
            - 'null'
          enum: [high, medium, low, null]
        evidence:
          type:
            - string
            - 'null'
        notes:
          type:
            - string
            - 'null'
        scheduledAt:
          type:
            - string
            - 'null'
          format: date-time
        performedAt:
          type:
            - string
            - 'null'
          format: date-time
        reportReceivedAt:
          type:
            - string
            - 'null'
          format: date-time
        outcomeNote:
          type:
            - string
            - 'null'
...
    ProstheticPrescriptionCreateRequest:
      type: object
      additionalProperties: false
      required:
        - patientId
        - prescribedAt
        - description
      properties:
        id:
          type: string
        patientId:
          type: string
        prescribedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, prescribed, submitted, authorized, delivered, tested, cancelled]
        category:
          type: string
          enum: [standard, oxygen, repair, replacement, trial, other]
        isoCode:
          type:
            - string
            - 'null'
        description:
          type: string
        measures:
          type:
            - string
            - 'null'
        clinicalReason:
          type:
            - string
            - 'null'
        regionalPrescriptionId:
          type:
            - string
            - 'null'
        supplier:
          type:
            - string
            - 'null'
        collaudoAt:
          type:
            - string
            - 'null'
          format: date-time
        collaudoOutcome:
          type:
            - string
            - 'null'
        source:
          type: string
          enum: [manual, document_review]
        documentRefs:
          type:
            - string
            - 'null'
        notes:
          type:
            - string
            - 'null'
...